Issue #2 Change SDK version in global.json to 3.1.200, to fix errors regarding different package versions Issue #2 Change Id type in Schedule entity from int to long Issue #2 Fix copy leftover mistake (Table property name from classrooms to terms) in Term entity Issue #2 Add InitialMigration to be used to create the database

+++ SSA2020-Back-Hypnotized-Chicken/SSA2020-Back-Hypnotized-Chicken.Data/Entities/Schedule.cs
@@ -4,11 +4,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace SSA2020_Back_Hypnotized_Chicken.Data.Entities
 {
 	[Table("schedules")]
-	public class Schedule : TimestampedEntity<int>
+	public class Schedule : TimestampedEntity<long>
 	{
 		[Key]
 		[Column("id", Order = 0)]
-		public override int Id { get; set; }
+		public override long Id { get; set; }
 		
 		[Required]
 		[Column("department", Order = 1)]
